team_management/requirements.txt
Django==1.11
django-rest-framework==0.1.0
djangorestframework==3.7.3
mysqlclient==1.3.12
pytz==2017.3
wheel==0.24.0
